+++ src/Control/Lens/Wrapped.hs
@@ -230,6 +230,12 @@ instance Wrapped (Last a) where
   _Wrapped' = iso getLast Last
   {-# INLINE _Wrapped' #-}
 
+instance (t ~ Alt g b) => Rewrapped (Alt f a) t
+instance Wrapped (Alt f a) where
+  type Unwrapped (Alt f a) = f a
+  _Wrapped' = iso getAlt Alt
+  {-# INLINE _Wrapped' #-}
+
 instance (t ~ ArrowMonad m' a', ArrowApply m, ArrowApply m') => Rewrapped (ArrowMonad m a) t
 instance ArrowApply m => Wrapped (ArrowMonad m a) where
   type Unwrapped (ArrowMonad m a) = m () a